refactor: move WebPushEventDocumentProps and WebPushEvent
