+++ Doppler.PushContact.Models/Doppler.PushContact.Models.csproj
@@ -6,4 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="MongoDB.Bson" Version="2.27.0" />
+  </ItemGroup>
+
 </Project>
