+++ Doppler.PushContact.WebPushSender.Test/Senders/DefaultWebPushSenderTest.cs
@@ -1,9 +1,9 @@
 using AutoFixture;
+using Doppler.PushContact.Models.Entities;
 using Doppler.PushContact.Models.Enums;
 using Doppler.PushContact.QueuingService.MessageQueueBroker;
 using Doppler.PushContact.WebPushSender.DTOs;
 using Doppler.PushContact.WebPushSender.Repositories.Interfaces;
-using Doppler.PushContact.WebPushSender.Repositories.Models;
 using Doppler.PushContact.WebPushSender.Senders;
 using Doppler.PushContact.WebPushSender.Test.Senders.Dummies;
 using Microsoft.Extensions.Logging;
