--- Doppler.PushContact.WebPushSender/Repositories/WebPushEventDocumentProps.cs
@@ -1,11 +1,0 @@
-namespace Doppler.PushContact.WebPushSender.Repositories
-{
-    public static class WebPushEventDocumentProps
-    {
-        public const string Id_PropName = "_id";
-        public const string Type_PropName = "type";
-        public const string MessageId_PropName = "message_id";
-        public const string PushContactId_PropName = "push_contact_id";
-        public const string Date_PropName = "date";
-    }
-}
